# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Epet API
  version: 0.1.0
  description: Draft MVP contract. PostgreSQL snapshots are authoritative; SSE only signals changes.
servers:
  - url: https://api.example.invalid
    description: Placeholder; real environment URLs are injected by deployment configuration.
...
paths:
...
  /v1/uploads:
    post:
      tags: [Uploads]
      operationId: createUpload
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadRequest'
      responses:
        '201':
          description: Restricted upload session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUploadResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /v1/uploads/{upload_id}/complete:
    post:
      tags: [Uploads]
      operationId: completeUpload
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UploadId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Object verification accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSnapshot'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/uploads/{upload_id}/content:
    put:
      tags: [Uploads]
      operationId: putUploadContent
      description: Local-development upload proxy. Production may instead return a restricted object-store URL.
      security: []
      parameters:
        - $ref: '#/components/parameters/UploadId'
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        '204':
          description: Bytes stored and integrity checked.
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/uploads/{upload_id}:
    delete:
      tags: [Uploads]
      operationId: deleteUpload
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UploadId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Deletion accepted.
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/generations:
    post:
      tags: [Generations]
      operationId: createGeneration
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGenerationRequest'
      responses:
        '201':
          description: Generation job created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationSnapshot'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
...
  /v1/models/{model_id}/download:
    post:
      tags: [Generations]
      operationId: downloadGenerationModel
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 128
      responses:
        '202':
          description: Model download command queued.
  /v1/models/{model_id}:
    delete:
      tags: [Generations]
      operationId: removeGenerationModel
      parameters:
        - name: model_id
          in: path
          required: true
          schema:
            type: string
            maxLength: 128
      responses:
        '202':
          description: Model removal command queued.
  /v1/generations/{job_id}:
    get:
      tags: [Generations]
      operationId: getGeneration
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Authoritative generation snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationSnapshot'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Generations]
      operationId: deleteGeneration
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Cloud deletion accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionSnapshot'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/deletions/{request_id}:
    get:
      tags: [Generations]
      operationId: getDeletion
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeletionId'
      responses:
        '200':
          description: Current deletion request status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionSnapshot'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/generations/{job_id}/events:
    get:
      tags: [Generations]
      operationId: streamGenerationEvents
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
            maxLength: 128
      responses:
        '200':
          description: Best-effort change notifications. Re-query snapshot on connect, gaps or reconnect.
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
...
  /v1/generations/{job_id}/artifact:
    get:
      tags: [Generations]
      operationId: getGenerationArtifact
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Short-lived download grant; only available in ready state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9_-]+$'
    UploadId:
      name: upload_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^upl_[A-Za-z0-9_-]{8,64}$'
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^gen_[A-Za-z0-9_-]{8,64}$'
    DeletionId:
      name: request_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^del_[A-Za-z0-9_-]{8,64}$'
...
  responses:
    Problem:
      description: Stable machine-readable error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    CreateUploadRequest:
      type: object
      additionalProperties: false
      required: [role, size, mime_type, sha256]
      properties:
        role:
          type: string
          enum: [primary, side, detail]
        size:
          type: integer
          minimum: 1
          maximum: 20971520
        mime_type:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
    CreateUploadResponse:
      type: object
      additionalProperties: false
      required: [upload_id, upload_url, allowed_headers, expires_at, limits]
      properties:
        upload_id:
          type: string
          pattern: '^upl_[A-Za-z0-9_-]{8,64}$'
        upload_url:
          type: string
          format: uri
        allowed_headers:
          type: object
          additionalProperties:
            type: string
        expires_at:
          type: string
          format: date-time
        limits:
          type: object
          additionalProperties: false
          required: [max_bytes]
          properties:
            max_bytes:
              type: integer
              minimum: 1
    UploadSnapshot:
      type: object
      additionalProperties: false
      required: [upload_id, status, version, updated_at]
      properties:
        upload_id:
          type: string
        status:
          type: string
          enum: [created, uploading, verifying, ready, failed, expired, deleted]
        version:
          type: integer
          minimum: 1
        updated_at:
          type: string
          format: date-time
        error:
          oneOf:
            - $ref: '#/components/schemas/ErrorDetail'
            - type: 'null'
    CreateGenerationRequest:
      type: object
      additionalProperties: false
      required: [primary_upload_id, additional_upload_ids, style_id, species, display_name]
      properties:
        primary_upload_id:
          type: string
          pattern: '^upl_[A-Za-z0-9_-]{8,64}$'
        additional_upload_ids:
          type: array
          maxItems: 2
          uniqueItems: true
          items:
            type: string
            pattern: '^upl_[A-Za-z0-9_-]{8,64}$'
        style_id:
          type: string
          minLength: 1
          maxLength: 64
        species:
          type: string
          enum: [cat, human]
        subject_kind:
          type: string
          enum: [pet_cat, human_avatar]
        display_name:
          type: string
          minLength: 1
          maxLength: 64
        provider_mode:
          type: string
          enum: [configured, auto, manual]
          default: configured
        requested_provider:
          type: string
          enum: [mock, cuda, openvino-gpu, openvino-cpu]
        requested_device_id:
          type: string
          maxLength: 128
    GenerationSnapshot:
      type: object
      additionalProperties: false
      required: [job_id, version, stage, retryable, created_at, updated_at]
      properties:
        job_id:
          type: string
          pattern: '^gen_[A-Za-z0-9_-]{8,64}$'
        version:
          type: integer
          minimum: 1
        stage:
          $ref: '#/components/schemas/GenerationStage'
        retryable:
          type: boolean
        progress:
          type: number
          minimum: 0
          maximum: 1
        provider:
          $ref: '#/components/schemas/GenerationProviderExecution'
        portrait_preview_url:
          type: string
          format: uri
        confirmation_expires_at:
          type: string
          format: date-time
        error:
          oneOf:
            - $ref: '#/components/schemas/ErrorDetail'
            - type: 'null'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GenerationProviderExecution:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode:
          type: string
          enum: [configured, auto, manual]
        requested:
          type: [string, 'null']
        device_requested:
          type: [string, 'null']
        actual:
          type: [string, 'null']
        device_actual:
          type: [string, 'null']
        model_id:
          type: [string, 'null']
        estimated_speed:
          type: [string, 'null']
...
    GenerationStage:
      type: string
      enum: [created, validating, generating_portrait, awaiting_portrait_confirmation, generating_actions, postprocessing, quality_check, packaging, ready, cancel_requested, failed, canceled, expired]
...
    ArtifactResponse:
      type: object
      additionalProperties: false
      required: [download_url, size, sha256, expires_at]
      properties:
        download_url:
          type: string
          format: uri
        size:
          type: integer
          minimum: 1
          maximum: 31457280
        sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
        expires_at:
          type: string
          format: date-time
    DeletionSnapshot:
      type: object
      additionalProperties: false
      required: [request_id, status, requested_at]
      properties:
        request_id:
          type: string
          pattern: '^del_[A-Za-z0-9_-]{8,64}$'
        status:
          type: string
          enum: [requested, processing, completed, failed]
        requested_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, params]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{2,63}$'
        params:
          type: object
          additionalProperties:
            type: [string, number, boolean]
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, trace_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          maxLength: 128
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{2,63}$'
        params:
          type: object
          additionalProperties:
            type: [string, number, boolean]
        trace_id:
          type: string
          minLength: 8
          maxLength: 128
security:
  - bearerAuth: []